Split up lib.diagnoseExpression into helper
